+++ drlpm/base/data_processing/indices/index_loader.py
@@ -1,7 +1,7 @@
 import logging
 import pandas as pd
 from drlpm.utils.logger import Logger
-from drlpm.data_processing.stocks.scraper_yahoo import ScraperYahoo
+from drlpm.base.data_processing.stocks.scraper_yahoo import ScraperYahoo
 
 
 class IndexLoader:
